refactoring -> use static qasm + dynamic .inc instead of autoqasm
